Implement a retry logic in Selenium recovery.
Selenium will sometime raise an exception (often a timeout) and fail.
We already detected this state and reinitialized Selenium.
But the client had to retry the request.

If this happened to FanFicFare during the metadata update, FFF would
not retry the request and just ignore the fiction.

Now, we will answer the client by sending the right content or an error
message.